fixed format modifier

+++ src/PrimaryVertexAnalyzer4PU.cc
@@ -1995,13 +1995,13 @@ PrimaryVertexAnalyzer4PU::analyze(const Event& iEvent, const EventSetup& iSetup)
       Fill(hnoBS,"yrecBeamvsdyXBS",recVtxs->begin()->yError(),recVtxs->begin()->y()-vertexBeamSpot_.y0());
 
       if(printXBS_) {
-	cout << Form("XBS %10d %5d %10d  %4d   %4lu %5.2f    %8f %8f       %8f %8f      %8f %8f",
-		     run_,luminosityBlock_,event_,bunchCrossing_,
-      		   recVtxs->begin()->tracksSize(), recVtxs->begin()->ndof(),
-      		   recVtxs->begin()->x(), 		   recVtxs->begin()->xError(), 
-      		   recVtxs->begin()->y(), 		   recVtxs->begin()->yError(), 
-      		   recVtxs->begin()->z(), 		   recVtxs->begin()->zError()
-      		   ) << endl; 
+	cout << Form("XBS %10d %5d %10d  %4d   %lu %5.2f    %8f %8f       %8f %8f      %8f %8f",
+		      run_,luminosityBlock_,event_,bunchCrossing_,
+		      (unsigned long)(recVtxs->begin()->tracksSize()), recVtxs->begin()->ndof(),
+		      recVtxs->begin()->x(), 		   recVtxs->begin()->xError(), 
+		      recVtxs->begin()->y(), 		   recVtxs->begin()->yError(), 
+		      recVtxs->begin()->z(), 		   recVtxs->begin()->zError()
+		     ) << endl; 
       }
 
     }
